pkg/river: introduce parser package (#1858)
* pkg/river: introduce parser package

This commit introduces the parser for River (grafana/agent#1839). It is
not wired in anywhere, and is the second of several PRs to slowly bring
over functionality from the prototype. The parser code has been highly
inspired by go/parser, with tests copied from the go/parser tests.

Note that this commit does not indicate we will ship river, just that we
are continuing the prototyping process in the main branch. pkg/river may
still be removed in the future in favor of an alternative solution for
Flow.

* update golangci-lint

* river: fix bugs found by parser fuzzer

* update Go and golanglint-ci versions in Drone for Windows

* update SHA for golangci-lint

* windows ci: only update Go since the Windows job doesn't do any linting

* .gitattributes: disable autocrlf and force \n line endings for fuzz tests

* wip: experiment with using a golang windows container for windows tests

* Revert "wip: experiment with using a golang windows container for windows tests"

This reverts commit a32b59d5cde485a78ba9881953ed58019ee76a99.

* river/parser: include comments when parsing

* river/parser: add missing ) in error message