river: add river/printer package (#1881)
* river: add river/printer package

river/printer is a pretty-printer of River AST nodes, which will be used
for error message formatting when a specific AST node should be
displayed to the user.

Much of the implementation river/printer is inspired by go/printer, and
so the way it formats files is very similar. For example, this input:

  secure = true // Enable secure mode
  access_log = "/tmp/access_log.txt" // Enable access log

is transformed to be aligned with elastic tabstops:

  secure     = true                  // Enable secure mode
  access_log = "/tmp/access_log.txt" // Enable access log

The formatting rules applied by river/printer are purposefully not
formally specified to allow them to change over time as formatting
recommendations are established.

river/printer will only print out files with LF; CRLF output is not
supported.

Additionally, since river/printer pretty-prints nodes, it also can serve
as the basis for a formatter. While a user-facing tool like `riverfmt`
is possible, one isn't introduced in this commit; more tests around
river/printer are needed for validating that the printer can never
produce a different AST than the one it formatted.

See pkg/river/printer/testdata for before-and-after formatting examples,
where *.in denote files prior to formating and *.expect denote what
files should look like after formatting.

* river/printer: add .gitattributes to force LF in testdata

* river/ast: consistently use *Pos naming for position fields