actions/stale: set close-issue-reason to not_planned
actions/stale just released v5.1.0 which includes the ability to set a
close reason for stale issues. Setting it to `not_planned` will mark
issues as [not-planned][] and can be filtered for when searching through
issues.

I don't think we need to update the uses: field of the action, and it'll
use the latest v5.Y.Z release.

[not-planned]: https://github.blog/changelog/2022-05-19-the-new-github-issues-may-19th-update/