river/internal/value: fix the representation of labeled blocks
This commit updates the reprsentation of blocks to be deeply nested
objects, with one key per name fragment in the name.

This means blocks named `a.b.c` will be represented as:

  {
    a = {
      b = {
        c = { /* block contents */ },
      },
    },
  }

If a block is labeled, its label is introduced as a top-level key. For
the block

  some_block "label" { }

it is represented as:

  {
    some_block = {
      label = { /* block contents */ }
    }
  }

This commit does not change the representation of slices of labeled
blocks, which will be updated separately.