river/internal/value: allow capsule markers and custom capsule conversions
Define Capsule interface which explicitly marks a Go type as being a
capsule. This is needed for Go types which should be capsules but whose
underlying types are non-capsules by default. This is required for types
like a Secret type, which would normally just be a string:

  type Secret string

  // RiverCapsule ensures that Secret values are treated as a capsule.
  func (s Secret) RiverCapsule() {}

ConvertibleCapsule is an extension interface to allow converting capsule
values from and to other Go values. For a Secret type, this will be used
to allow converting a string into a Secret, but not a Secret into a
string.