River: capsule markers, capsule conversion, and custom tokenization  (#1933)
* river/internal/value: allow capsule markers and custom capsule conversions

Define Capsule interface which explicitly marks a Go type as being a
capsule. This is needed for Go types which should be capsules but whose
underlying types are non-capsules by default. This is required for types
like a Secret type, which would normally just be a string:

  type Secret string

  // RiverCapsule ensures that Secret values are treated as a capsule.
  func (s Secret) RiverCapsule() {}

ConvertibleCapsule is an extension interface to allow converting capsule
values from and to other Go values. For a Secret type, this will be used
to allow converting a string into a Secret, but not a Secret into a
string.

* river/token/builder: allow custom tokenizers for printing values

Define a Tokenizer interface which is used to print custom tokens for a
Go value. This will allow custom printing of types like an
OptionalSecret, which should be shown as a literal string when it is
marked as cleartext, and hidden when it is marked as sensitive.

* river/internal/value: split ConvertibleCapsule into two interfaces

Split ConvertibleCapsule into two separate ConvertibleFromCapsule and
ConvertibleIntoCapsule types. This allows for non-pointer capsule values
to be used for converting into another pointer, which is not possible
with the combined interface (since ConvertInto must be a pointer
method).

* use a struct to ensure consistent ordering of object keys

* river/internal/value: remove flaky tests

Testing conversion from map[string]interface{} to an object is likely to
fail due to inconsistent key orderings while iterating over a map.