river/ast: create ast.Walk function
ast.Walk can be used to walk an AST. This will be used by Flow to locate
references to other components.

As a part of this change, *ast.Ident has been added as a node to be able
to distinguish when an identifier is used as an expression or as part of
a node.