river/internal/value: minimize usage of Value.rv in decoder
It is generally incorrect to rely on Value.rv in the decoder, since the
underlying Go type can change. The Decoder should instead rely on the
exposed API.

This doesn't completely eliminate usage of Value.rv, but takes care of
easy-to-fix cases where it was being used unnecessarily. Further commits
to completely eliminate the usage of it will be done in the future.