River: various small fixes (#1951)
* river/internal/value: support new Unmarshaler interface

* river/diag: initial commit of river/diag package

river/diag is a new package which contains common error types and a
printer which can format them in a user friendly way, including the
contents from the files where the error originated.

This package is then used consistently from other packages instead of
each package defining its own error type.

* river/ast: create ast.Walk function

ast.Walk can be used to walk an AST. This will be used by Flow to locate
references to other components.

As a part of this change, *ast.Ident has been added as a node to be able
to distinguish when an identifier is used as an expression or as part of
a node.

* river/builder/token: write out comments properly

* river/token/builder: allow encoding bodies from go values

This emulates the gohcl.EncodeIntoBody functionality needed to write the
/-/config endpoint completely.

* river/internal/value: minimize usage of Value.rv in decoder

It is generally incorrect to rely on Value.rv in the decoder, since the
underlying Go type can change. The Decoder should instead rely on the
exposed API.

This doesn't completely eliminate usage of Value.rv, but takes care of
easy-to-fix cases where it was being used unnecessarily. Further commits
to completely eliminate the usage of it will be done in the future.

* river/internal/value: make Value aware of TextMarshaler/time.Duration

It was incorrect to add logic to token/builder for handling the
interfaces, since the Go values themselves were never treated as
strings, which is the intended behavior.

* river/internal/value: properly detect River type of addressable values

This fixes an issue where the River type of an addressable value was
improperly detected, as value.makeValue fully deferences types before
detecting the River type.

This meant that types which only implement encoding.TextMarshaler for a
pointer receiver (i.e., model.Duration) weren't properly be detected as
River strings.

Now, type assignment is always properly done with addressable values.

* go mod tidy

* fix lint