river/scanner: scan for single quoted strings to reject explicitly
This explicitly detects single quoted strings to return as a single
illegal token error, rather than a bunch of different illegal tokens.

This restriction may be lifted in the future and have single quoted
strings be permitted, but is put in place for now so at least the error
messages are nicer for users who expect single quoted strings to work.