operations/river-jsonnet: allow river.expr() to be used with std.prune()
std.prune removes all hidden fields. This commit changes River exprs to
have a non-hidden field which indicates that it's an expression.

The new field is constructed with a name which could never appear as an
attribute in a River block, as River attributes can only be valid
identifiers, and identifiers cannot start with `$`.