Introduce Makefile.v2
This commit temporarily introduces a Makefile.v2, introduced initially
as a separate file for easier comparison to the existing Makefile.

Makefile.v2 is intended to eventually replace the original Makefile and
follows the designed outlined in grafana/agent#2052.

It includes targets only needed for local development and testing.
Targets only intended for CI are purposefully omitted to keep the
Makefile as simple as possible.

It is controlled primarily by environment variables, which are
documented at the comment at the top of the Makefile. That comment may
be printed to stdout by calling `make -f Makefile.v2 help`.

The new build image, grafana/agent-build-image:v0.15.0, is used for
proxying Makefile commands when the `USE_CONTAINER=1` environment
variable is set.

A Docker volume cache can be initialized to speed up builds in the build
container. Run `make -f Makefile.v2 build-container-cache` to create the
cache and `make -f Makefile.v2 clean` to remove it. The cache stores the
Go module cache and Go build cache. The Makefile will detect the cache
and automatically use it when proxying commands to the build container.
Credit to @captncraig for the idea of using Docker volumes for caching
builds.

This first commit is opened early for feedback on the approach. Missing
functionality will be added in future commits if the approach seems OK:

1. There is currently no equivalent target to test-packages
2. The Dockerfiles still build the agent instead of copying over a built
   binary