add new Dockerfiles
Add new Dockerfiles which use Makefile-v2.mk for building the image and
caches for running the make command.

This allows the Dockerfile to continue building the image while taking
advantage of a shared Go module and Go build cache for improved
subsequent build times.

The Dockerfile.new files are added as new files to make comparison
easier. They will fully replace the Dockerfile and Dockerfile.buildx
variants once Makefile-v2 replaces the existing Makefile.

This does not support building multiplatform images, which is only
useful from the context of CI (as producing multiplatform images
requires pushing the images to the registry, which should never be done
locally during development)