prometheus.remote_write: finalize prometheus.remote_write component
The following changes are made to prometheus.remote_write:

* Naming patterns are changed to align with other components. For
  example, using an Arguments struct to describe the arguments of the
  component.

* The nested `remote_write` block has been renamed to `endpoint`.

* Missing options are added for the `endpoint` block:
  * `remote_timeout`
  * `headers`
  * `metadata_config`

* `client_config` is renamed to `http_client_config` to align with other
  components.

* The `basic_auth` block is removed, as that block is already defined in
  `http_client_config`.

The remainder of the changes are minor code cleanup. A test has been
added to assert that prometheus.remote_write works as expected by
sending synthetic metrics to a test server.

The following has been purposefully scoped out of this PRs and will be
done in later PRs:

* Supporting the sigv4 authentication method in the `endpoint` block.
* Updating documentation to reflect the naming changes.