component/otelcol/exporter: initial commit
This commit introduces a new package, component/otelcol/exporter, which
exposes a generic Flow component implementation which can run
OpenTelemetry Collector exporters.

There is some stuff left to do for this implementation to be complete:

* A Zap logging adapter needs to be created to correctly process logs
  from OpenTelemetry Collector components.
* Component-specific metrics are currently ignored.
* Component-specific traces are currently ignored.

All of the above will be done in separate PRs.

As of this commit, there are no registered `otelcol.exporter.*`
components. Implementations for OpenTelemetry Collector Flow components
will be done in future PRs.

Related to #2213.