Flow: add otelcol.receiver.otlp component (#2326)
Introduce an otelcol.receiver.otlp component which can receive OTLP data
over gRPC or HTTP.

This is a nearly 1:1 match to the upstream component with one exception:
Custom authentication methods are not yet supported.

Future work will be needed to figure out how we want to expose
authentication to components. The way it currently works in
OpenTelemetry is by exposing an Authentication extension to components.
We will likely need to do something different. Work on this will be done
in a future PR.

Closes grafana/agent#2286.