Duplicate linux install script, remove fetching cloud config.
The cloud functionality to generate the agent config file is soon to be
removed, hence here we duplicate the linux install script so that both
versions can be used while the cloud change is being worked on.

The new script is identical to the old one, except for `grafana-agentctl
cloud-config` being removed (together with env vars for auth).