prometheus.*: immediately export receiver on component construction
This commit changes Flow Prometheus components which export receivers to
immediately export their receivers on component construction. Receivers
for these components remain the same through the process lifetime, so
there's no need for them to emit exports at run or update time.

This is also a minor (but likely immeasurable) performance enhancement,
where components depending on `prometheus.remote_write` or
`prometheus.relabel` won't be forced to update twice in a row near the
process start time.

The CHANGELOG has been updated to reflect the performance improvement
angle, since it is the only way for the changes to be observed by end
users.

Closes #1946.