river/encoding: fix issue where interface capsules were downcasted to empty objects
This change is similar to the change introduced in #2369, where calling
the `Interface()` method on River values forced a type change from an
interface-based capsule to an any-type capsule.

The new `value.FromRaw` method allows constructing a River value from a
reflect.Value directly, avoiding the need to convert to interface{}.

As part of this change, a `!reflect.Value.IsZero()` check has been added
for safety; this shouldn't happen in most circumstances unless a
value.Null is trying to be encoded.

Fixes #2361.