Use Go 1.19 for builds (#2455)
* update to Go 1.19.3

* Enable optimizations and inlining during tests

Passing `-gcflags "all=-N -l"` disables optimizations and inlining. This
is primarily useful for debuggers, and the original flag was inherited
through Loki (which was inherited through Cortex). However, it causes
issues in Go 1.19 (see issue 54291 in the Go repo).

This flag should no longer be necessary to manually pass in the Makefile
anymore, as tools like delve and GoLand should instead be resonsible for
passing it when debugging is initiated.

Re-enabling optimizations and inlining fixes the issue, and also serves
to make test code faster.