remote.http: new component
Introduce a `remote.http` component which polls a specified URL. The
response body of the request is exposed to other components as either a
string or a secret, depending on the value of the `is_secret` argument.

This component acts as a generic form of Prometheus' HTTP SD, so it
implements #2270.

While documenting this component, I noticed some of the shared blocks
accidentally included component-specific detail. All of the shared
blocks have been updated to make sure they're documenting the correct
information at a generic level.

Closes #2270.