Support exemplars in Prometheus component pipelines (#2574)
This commit allows exemplars to be passed around in Prometheus component
pipelines. To support this change, the prometheus.Interceptor struct was
extended to support intercepting any call to the Appendable interface.

Support for metadata is also added, though that code path is currently
not invoked in our usage of the Prometheus scraper.

Fixes #2420.