component/otelcol: expose metrics from upstream otelcol components
This commit configures the MeterProvider given to otelcol components to
forward metrics to a Prometheus Registry. A new Prometheus Registry is
generated every time otelcol components are created and passed to a lazy
collector to hot-swap out the Registry being used.

For this to work, the telemetry.useOtelForInternalMetrics feature gate
from OpenTelemetry Collector is forcibly enabled, which tells the
go.opentelemetry.io/collector/obsreport package (which is used by some
components for generating metrics) to use the provided MeterProvider
instead of OpenCensus for metrics.

Note that the OpenTelemetry Collector dependency will need to be updated
to v0.67 before we can make full use of this change, as v0.67 includes
changes to obsreport so that the MeterProvider is used for all component
types.

Closes #2372.