tools/agentlint: introduce custom linting (#2594)
* tools/agentlint: prototype custom linters

This commit prototypes a custom linter for ensuring that Flow components
are imported.

It will investigate `./component/all`, doing the following:

1. Retrieve the list of direct imports of `./component/all`.
2. Examine packages in `./component/...` to determine which declare a Flow
   component.
  1. A package declares a Flow component if it has an `init` function 
     which calls `component.Register`.
3. Ensure that `./component/all` imports all of the discovered component packages.

Run `make agentlint` to build the linter at `./build/agentlint`. 

To run it locally, run the following, and observe that no errors are reported:

```
agentlint ./...
```

You can remove an entry from `./component/all` and rerun to observe errors
being reported. 

The linter is hooked up to the `make lint` target, and we will be able to add
more custom linter passes in the future to agentlint as needed.

* pkg/util: add Eventually

util.Eventually assists with writing tests which rely on asynchronous
behavior.

It is currently not used.