component/common/loki: drop unqueued logs after 5 seconds on shutdown (#2721)
Fix an issue where being unable to send logs to `loki.write` due to the
client being permanently backlogged would deadlock the Flow controller.

The `loki.write` client may be permanently backlogged when:

* Limits are reached when sending logs to Loki, leading to endless
  request retries.
* Loki has an extended outage.

When an EntryHandler is stopped, it will wait for 5 seconds before
forcibly stopping the goroutine which queues log entries. If this
timeout is reached, any unqueued log entries are permanently lost, as
the positions file will likely be updated past the point where the entry
was read.

While losing logs is not ideal, it's unacceptable for any Flow component
to be able to block the controller. This is a short-term solution to
allow the Flow controller to continue working properly. A long term
solution would be to use a Write-Ahead Log (WAL) for log entries. See
grafana/loki#7993.

Fixes #2716.
Related to grafana/loki#2361.