Prep for v0.30.2 (#2723)
* prometheus.relabel: clone labels before relabeling (#2701)

This commit clones the label set before applying relabels. Not
cloning does two things:

1. It forces the computed ID of the incoming series to change (as its
   labels changed)

2. It can cause obscure bugs with relabel rules being applied, such as
   a `keep` action which doesn't work after modifying the original
   slice.

* component/common/loki: drop unqueued logs after 5 seconds on shutdown (#2721)

Fix an issue where being unable to send logs to `loki.write` due to the
client being permanently backlogged would deadlock the Flow controller.

The `loki.write` client may be permanently backlogged when:

* Limits are reached when sending logs to Loki, leading to endless
  request retries.
* Loki has an extended outage.

When an EntryHandler is stopped, it will wait for 5 seconds before
forcibly stopping the goroutine which queues log entries. If this
timeout is reached, any unqueued log entries are permanently lost, as
the positions file will likely be updated past the point where the entry
was read.

While losing logs is not ideal, it's unacceptable for any Flow component
to be able to block the controller. This is a short-term solution to
allow the Flow controller to continue working properly. A long term
solution would be to use a Write-Ahead Log (WAL) for log entries. See
grafana/loki#7993.

Fixes #2716.
Related to grafana/loki#2361.

* prepare for v0.30.2 release

* address review feedback

* operator: Use enableHttp2 field as boolean in libsonnet templates (#2724)

Signed-off-by: Paschalis Tsilias <paschalis.tsilias@grafana.com>

Signed-off-by: Paschalis Tsilias <paschalis.tsilias@grafana.com>
Co-authored-by: Paschalis Tsilias <tpaschalis@users.noreply.github.com>