Ensure server LogLevel and LogFormat have defaults (#2607)
Some YAML parsing of the Server configs will default
initialize LogLevel and LogFormat to blank.  This adds
a check and corrects the values to the defaults. Also
adds a unit test.

Other packages have `ApplyDefaults` methods to handle
complex initialization.  We follow this pattern and allow `config.Config`
to fix-up `server.Config` with correct defaults.