operations/helm: remove warning about Chart not being ready for use (#2735)
The warning about the chart being experimental was only meant for the
prototype before it got upstreamed, and was accidentally kept in.

Additionally, change ct.yaml to no longer require every change to have a
version bump. This allows us to batch changes without having to make a
release for every single one.

To publish the changes to the README and remove the warning, the version
has been bumped to 0.2.1.