pkg/runner: introduce common package for goroutine scheduling (#2788)
Many of our Flow components share a behavior of managing goroutines,
where they must synchronize the set of running goroutines with a set of
tasks to perform:

* `loki.source.file` must manage tailer goroutines for each logs target.
* `phalre.scrape` must manager scraper goroutines for each profiling target.
* `prometheus.scrape` must manage scraper goroutines for each metrics
  target.

More components with this type of goroutine management will be added in
the future (#2504, #2763). Even the Flow controller implements exactly
the same goroutine management to manage the set of running components.

Throughout all of these uses, the goroutine management works the same
way:

* Start new goroutines for tasks (e.g., targets) that have appeared.
* Shut down old goroutines for tasks that have gone away.
* Maintain existing goroutines as long as the task continues to exist.

This commit introduces a new utility package, `pkg/runner`, which
creates an abstraction for this behavior. `pkg/runner` defines three
concepts:

* Task: a unit of work that a goroutine should be performing, such as a
  metrics target.
* Worker: a goroutine which will act against a Task, such as scraping a
  metrics target on an interval.
* Runner: a mechanism that controls the set of running Workers based on
  Tasks given to it.

The Runner is a declarative interface; you call `ApplyTasks` with the
entire set of Tasks to run, and it will synchronize the set of Workers
with the given Tasks, shutting down Workers for Tasks that have gone
away, creating new Workers for new Tasks, and maintaining existing
Workers for Tasks which still exist.

The Task interface is defined as:

```go
type Task interface {
  Hash() uint64
  Equals(other Task) bool
}
```

Tasks are tracked internally in a hash map. The `Hash` method is invoked
to figure out what bucket to put a Task in. If two Tasks happen to have
the same hash, the `Equals` method is called to check if they are equal.
Completely identical tasks (with the same hash and where `Equals`
returns true) will be deduplicated so only one task exists.

Meanwhile, the Worker interface is defined as:

```go
type Worker interface {
  Run(ctx context.Context)
}
```

The `Run` method is invoked when a Worker should start running. Workers
should continue running until there is an error or the provided context
is canceled, which is done by the Runner when a Task assigned to a
Worker is removed.

There is exactly one Worker for each unique Task. Each Worker is
responsible for exactly one task for its entire lifetime. A Runner is
constructed with a function to map Tasks to Workers:

```go
func New[TaskType Task](newWorker func(t TaskType) Worker) *Runner[TaskType]
```

`newWorker` is only invoked when the Runner is given a Task for which no
Runner is assigned to. To provide Tasks to a runner, call `ApplyTasks`.

See included test code for example of how to use the `runner` package.

Benchmarks are included to test the hashMap, where the bulk of the
performance will be seen. Benchmark results show modest performance for
the initial commit:

    Benchmark_hashMap/Add_element-10                                 7875709           133.3  ns/op
    Benchmark_hashMap/Remove_element-10                             30147565            48.13 ns/op
    Benchmark_hashMap/Lookup_element-10                             85129270            14.20 ns/op
    Benchmark_hashMap/Add_existing_element-10                       61546616            18.85 ns/op
    Benchmark_hashMap_collisions/1_collisions/Add-10                 5402604           215.6  ns/op
    Benchmark_hashMap_collisions/1_collisions/Remove-10             32215772            37.19 ns/op
    Benchmark_hashMap_collisions/1_collisions/Lookup-10             40548360            29.54 ns/op
    Benchmark_hashMap_collisions/2_collisions/Add-10                 4416198           270.9  ns/op
    Benchmark_hashMap_collisions/2_collisions/Remove-10             21474396            55.78 ns/op
    Benchmark_hashMap_collisions/2_collisions/Lookup-10             26042961            45.52 ns/op
    Benchmark_hashMap_collisions/5_collisions/Add-10                 3013473           396.2  ns/op
    Benchmark_hashMap_collisions/5_collisions/Remove-10             10754502           111.6  ns/op
    Benchmark_hashMap_collisions/5_collisions/Lookup-10             11606706           102.8  ns/op
    Benchmark_hashMap_collisions/10_collisions/Add-10                1863262           644.7  ns/op
    Benchmark_hashMap_collisions/10_collisions/Remove-10             5846124           204.8  ns/op
    Benchmark_hashMap_collisions/10_collisions/Lookup-10             5014551           241.2  ns/op
    Benchmark_hashMap_collisions/100_collisions/Add-10                 81848         14595    ns/op
    Benchmark_hashMap_collisions/100_collisions/Remove-10             635085          1877    ns/op
    Benchmark_hashMap_collisions/100_collisions/Lookup-10              90588         13244    ns/op

As expected, the performance steadily decreases when more hash
collisions are introduced. Proper usage of pkg/runner will be to make
sure that there is good hash coverage through different tasks.