loki.source.kubernetes: new component
The `loki.source.kubernetes` component, when given a set of targets from
`discovery.kubernetes`, tails logs from containers using the Kubernetes
API.

One tailer is launched for each unique container. Tailers will reconnect
to Kubernetes with exponential backoff if their connection to the
Kubernetes API is closed. Tailers only exit if the target is removed
from the input or if the tailer determines that the terminated container
will not restart.

Closes #2504.