Flow: add loki.source.kubernetes component (#2817)
* pkg/runner: add Tasks method to return tracked tasks

* pkg/runner: don't delete workers when the worker exits automatically

This allows Tasks() to return all scheduled tasks, even if the runner
exited. This is useful for showing DebugInfo in components regardless if
the runner is actively running.

* component/discovery: add Labels utility method to discovery.Target

Labels will convert a discovery.Target into Prometheus' labels.Labels
type.

* loki.source.kubernetes: new component

The `loki.source.kubernetes` component, when given a set of targets from
`discovery.kubernetes`, tails logs from containers using the Kubernetes
API.

One tailer is launched for each unique container. Tailers will reconnect
to Kubernetes with exponential backoff if their connection to the
Kubernetes API is closed. Tailers only exit if the target is removed
from the input or if the tailer determines that the terminated container
will not restart.

Closes #2504.