Use organization ISSUE_COMMANDS_TOKEN with reduced scope
The GH_BOT_ACCESS_TOKEN has uses beyond the addToProject and other API
commands that the issue_commands workflow uses.

The new token is set at an organization level so it does not require
repository administrators to rotate the token. It also has the minimal
classic PAT permissions to facilitate the workflow.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>