rivertags: support enum fields
An enum field is a slice of blocks where each element of the slice will
only have one block set inside of it.

Enum fields must have a common base name which will be appended to the
name of blocks inside the slice. For example, if the base name is
`stage`, and a block inside the enum slice is `json`, then the full
block name is `stage.json`.