Ensure empty server config fails to start (#2607)
Some YAML parsing of the Server configs will default
initialize LogLevel and LogFormat to blank.  These are
invalid values causing the agent to log verbosely.

This commit changes the `config.Config.Server` member to a
pointer type.  When config marshalling results in a nil `Server`,
an error is returned.   The agent will fail out alerting the user
of the issue, and thus is not running in an invalid state.

Also adds a unit test `TestConfig_EmptyServerConfigFails`
which specifically exercises the "load from file" code
path and ensures an invalid empty server returns an error.