flow: support a controller ID
This commit adds support for a "controller ID." This ID is used to
generally globally unique names for components.

It is required that, for the scope of a process, the following are
unique per Flow controller:

* ControllerID
* DataPath
* HTTPPath

Component IDs can be thought of having:

* Local ID: The ID of the component as denoted by the config file.
* Global ID: The combination of the controller ID and the local ID.

The global ID is used for telemetry data and is given to components for
the purposes of telemetry data generation.

Because DataPath and HTTP path must be globally unique, only the local
ID is used when generating data directories or HTTP paths for
components.