river: allow referencing zero values in Go maps
This fixes an issue where any reference to a zero value in a Go map
would be treated as if the key didn't exist. This was due to a
misunderstanding of the Go documentation that said "the zero Value" will
be returned if the key does not exist in the map. This documentation
referred to an invalid `reflect.Value` instead of the zero value of the
map's value type.

This allows to simplify the logic when performing a map lookup to
determine if the key was found or not, and permits expressions which
reference a map key set to the zero value of the value type.