Flow: support modules through isolated controllers (#3179)
* flow: support HTTP path prefixes for components

Add support for a Flow controller being given an HTTP path prefix to
propagate to components. This is necessary when the Flow controller is
exposed via a non-root path.

This then moves the current path prefix to the root binary.

* flow: support a controller ID

This commit adds support for a "controller ID." This ID is used to
generally globally unique names for components.

It is required that, for the scope of a process, the following are
unique per Flow controller:

* ControllerID
* DataPath
* HTTPPath

Component IDs can be thought of having:

* Local ID: The ID of the component as denoted by the config file.
* Global ID: The combination of the controller ID and the local ID.

The global ID is used for telemetry data and is given to components for
the purposes of telemetry data generation.

Because DataPath and HTTP path must be globally unique, only the local
ID is used when generating data directories or HTTP paths for
components.

* flow: add concept of controller arguments

Supported controller arguments are determined by the number of
`arguments` block provided to the config file. Argument values are
supplied by passing a map to flow.LoadFile. These are then exposed to
components through a vm.Scope.

* flow: add concept of controller exports

Controller exports are exposed by adding an `export` block to the config
file, where the value attribute determines what that export will
evaluate to. After evaluating exports, a callback is called to inform
the controller's owner.

* grafana-agent: change base HTTP prefix to /api/v0/component/

The original /component/ prefix collided with the UI, and broke when
retrieving a deeply nested path.

* web/api: move /api/v0/web prefix to main

This allows to use the Flow web API with different prefixes, such as
from the scope of a module loader.

* web/ui: add modules support

This change makes the UI modules-aware, where viewing a module loader
page will show the components running inside of that module.

The UI for components now supports multiple paths, where each /
indicates a component running inside of a module.