river/vm: support decoding blocks into map[string]T (#3232)
This allows decoding blocks into maps instead of structs. The map type
can be any map[string]T, which is useful of the map values must always
be a consistent type. If map values can be anything, use map[string]any
for decoding.

Additionally, if decoding into an interface{}, it will default to
map[string]any when decoding.

When decoding into a map, nested blocks aren't allowed and will be
rejected.

Closes #2903.