Fix ambiguous link (#3233)
When a relref target is not explicitly set as relative to the current
page with a preceding `./` or `../`, Hugo attempts to resolve the
relref target by seeing if it could refer any page in the site.

For local builds, this relref results in the relative permalink
/docs/agent/latest/flow/concepts/configuration_language/#Attributes
because of the docs/sources/flow/concepts/configuration_language.md
file. This works locally because there is only one file with this
name. However, in the website, multiple files have this name because
of the multiple published versions of Agent documentation. When the
resolution is ambiguous, Hugo prefers to avoid linking to any of the
pages and instead makes the permalink point to the current page.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>
(cherry picked from commit b8b36d35c481ca468122f3147fad9637c8adc550)