flow/tracing: fix issue where remote sampler could not be parsed (#3240)
This fixes an issue where the remote sampler could not be parsed if it
used strings for the strategy type instead of numbers. The resolution is
to switch to the protobuf JSON encoding, which properly parses this type
as a string.

Not being able to parse the remote sampler rules meant that sampling
rate fell back to the default (10%).

This change is equivalent to
open-telemetry/opentelemetry-go-contrib#3183. We are still unable to
move to the upstream type while
open-telemetry/opentelemetry-go-contrib#2981 remains unresolved.