fix script zip expand issues
Prior to changes here, the Expand-Archive command was extracting the agent installer into a path with the same name and then not executing it from the path causing the script to fail.  These changes get the current path and use that in the -DestinationPath parameter, effectively extracting the ZIP into the same directory the ps1 script runs in which is also where the zip is downloaded to. This fixes the installation failing.