flow/logging: introduce prototype v2 logging subsystem
This commit proposes a solution for grafana/agent#3107 by creating a new
logging subsystem composed of:

* A logging `Sink`: a place to send emitted logs
* A `Logger`: a `go-kit/log.Logger` implementation which sends logs to a
  configured sink.

Compared to the current logging subsystem, this breaks apart
logging.Logger into the Sink and Logger pairs. In this new API, the Sink
is what supports being updated with new Options.

Additionally, Loggers may now be constructed with options, including
assigning a Component ID.

There are two different types of sinks:

* A `WriterSink` formats and writes logs to an `io.Writer`.
* A `LoggerSink` forwards logs to another `*Logger` instance.

When `LoggerSink` is used, the component ID of the Logger (if any)
propagates properply to loggers generated with that sink, ensuring a
global ID is generated.

The intended usage is that flow.Options is updated to be given a Sink
rather than a logging subsystem, and flow.Options will create a Logger
for itself, and create a set of Loggers for managed components.

Components which create their own controllers (like module.string from
grafana/agent#3227) will create a new sink from the instance of
`*logging.Logger` given to them.

See the included example code for how this may be used in practice.