flow/logging: rewrite logging subsystem (#3259)
* flow/logging: rewrite logging subsystem

This commit rewrites the Flow logging subsystem so it can be used as part of
the solution for #3107. 

The new logging subsystem is composed of:

* A logging `Sink`: a place to send emitted logs.
* A `Logger`: a `go-kit/log.Logger` implementation which sends logs to a
  configured sink.

Compared to the previous logging subsystem, this breaks apart logging.Logger
into the Sink and Logger pairs. Now, the Sink is what supports being updated
with new Options.

Additionally, Loggers may now be constructed with options, including assigning
a Component ID.

There are two different types of sinks:

* A `WriterSink` formats and writes logs to an `io.Writer`.
* A `LoggerSink` forwards logs to another `*Logger` instance.

When `LoggerSink` is used, the component ID of the Logger (if any) propagates
properply to loggers generated with that sink, ensuring a global ID is
generated.

The Flow controller and components have been updated to use this API explicitly
to handle part of grafana/agent#3107.

* flow: create Flow-specific stdlib

This decouples Flow-specific stdlib functions to avoid import cycles.